+++ lib/new_relic/rack/browser_monitoring.rb
@@ -111,7 +111,6 @@ module NewRelic
       end
 
       def html?(headers)
-        # needs else branch coverage
         headers[CONTENT_TYPE]&.match?(TEXT_HTML)
       end
 
@@ -154,7 +153,7 @@ module NewRelic
 
       def gather_source(response)
         source = nil
-        response.each { |fragment| source ? (source << fragment.to_s) : (source = fragment.to_s) }
+        response.each { |fragment| source ? (source << fragment.to_s) : (source = fragment.to_s.dup) }
         source
       end